webservice/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.17.0
    # via tortoise-orm
anyio==3.6.1
    # via starlette
async-timeout==4.0.3
    # via aiohttp
attrs==22.1.0
    # via aiohttp
boto3==1.34.19
    # via -r requirements.in
botocore==1.34.19
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via uvicorn
colorama==0.4.5
    # via
    #   click
    #   loguru
fastapi==0.94.0
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.13.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
    #   yarl
iso8601==1.1.0
    # via tortoise-orm
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
loguru==0.6.0
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.23.3
    # via
    #   -r requirements.in
    #   pandas
orjson==3.9.10
    # via -r requirements.in
packaging==21.3
    # via setuptools-scm
pandas==1.4.4
    # via -r requirements.in
pydantic==1.9.1
    # via
    #   -r requirements.in
    #   fastapi
pyhumps==3.8.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pypika-tortoise==0.1.6
    # via tortoise-orm
python-dateutil==2.8.2
    # via
    #   botocore
    #   pandas
pytz==2022.2.1
    # via
    #   -r requirements.in
    #   pandas
    #   tortoise-orm
pyyaml==6.0
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
s3transfer==0.10.0
    # via boto3
setuptools-scm==6.4.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.2.0
    # via anyio
starlette==0.26.0.post1
    # via fastapi
tomli==2.0.1
    # via setuptools-scm
tortoise-orm==0.19.3
    # via -r requirements.in
typing-extensions==4.3.0
    # via
    #   aiosqlite
    #   pydantic
urllib3==1.26.14
    # via
    #   botocore
    #   requests
uvicorn==0.18.2
    # via -r requirements.in
win32-setctime==1.1.0
    # via loguru
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
